# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Urban Up - Smart City & Tourism Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /traffic/route:
    post:
      operationId: getSmartRoute
      tags: [traffic]
      summary: Get AI smart routing with alternative routes and point rewards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RouteRequest"
      responses:
        "200":
          description: Smart route response with alternatives and point rewards
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SmartRouteResponse"
...
  /accessibility/obstacles:
    get:
      operationId: getObstacles
      tags: [accessibility]
      summary: Get reported obstacles on safe paths
      responses:
        "200":
          description: List of obstacles
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Obstacle"
    post:
      operationId: reportObstacle
      tags: [accessibility]
      summary: Report a dynamic obstacle on a safe path
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ObstacleInput"
      responses:
        "201":
          description: Obstacle reported
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Obstacle"
...
  /points/redeem:
    post:
      operationId: redeemPoints
      tags: [points]
      summary: Redeem Jawwal Points for rewards
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RedeemInput"
      responses:
        "200":
          description: Redemption successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointsTransaction"
...
  /auth/user:
    get:
      tags: [Auth]
      operationId: getCurrentAuthUser
      summary: Get the currently authenticated user
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
      responses:
        '200':
          description: Auth status resolved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthUserEnvelope'
  /login:
    get:
      tags: [Auth]
      operationId: beginBrowserLogin
      summary: Start the browser OIDC login flow
      parameters:
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the OIDC authorization endpoint.
  /callback:
    get:
      tags: [Auth]
      operationId: handleBrowserLoginCallback
      summary: Complete the browser OIDC login flow
      parameters:
        - name: code
          in: query
          required: false
          schema:
            type: string
        - name: state
          in: query
          required: false
          schema:
            type: string
        - name: iss
          in: query
          required: false
          schema:
            type: string
      responses:
        '302':
          description: Redirect to the returnTo path on success or /api/login on failure.
  /logout:
    get:
      tags: [Auth]
      operationId: logoutBrowserSession
      summary: Clear the session and begin OIDC logout
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
        - $ref: '#/components/parameters/SessionCookie'
        - name: returnTo
          in: query
          required: false
          schema:
            type: string
            default: /
      responses:
        '302':
          description: Redirect to OIDC provider logout URL.
  /mobile-auth/token-exchange:
    post:
      tags: [Auth]
      operationId: exchangeMobileAuthorizationCode
      summary: Exchange a mobile OIDC code for a session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [code, code_verifier, redirect_uri, state]
              properties:
                code:
                  type: string
                  minLength: 1
                code_verifier:
                  type: string
                  minLength: 1
                redirect_uri:
                  type: string
                  minLength: 1
                state:
                  type: string
                  minLength: 1
                nonce:
                  type: string
      responses:
        '200':
          description: Session created successfully.
          content:
            application/json:
              schema:
                type: object
                required: [token]
                properties:
                  token:
                    type: string
        '400':
          description: Missing or invalid fields.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '401':
          description: Token did not contain usable claims.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Token exchange failed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /mobile-auth/logout:
    post:
      tags: [Auth]
      operationId: logoutMobileSession
      summary: Delete a mobile session token
      parameters:
        - $ref: '#/components/parameters/AuthorizationSessionHeader'
      responses:
        '200':
          description: Logout processed.
          content:
            application/json:
              schema:
                type: object
                required: [success]
                properties:
                  success:
                    type: boolean
components:
  parameters:
    AuthorizationSessionHeader:
      name: Authorization
      in: header
      required: false
      description: "Opaque session token — Bearer <sid>."
      schema:
        type: string
    SessionCookie:
      name: sid
      in: cookie
      required: false
      description: Browser session cookie.
      schema:
        type: string
  schemas:
...
    RouteRequest:
      type: object
      required: [originLat, originLng, destLat, destLng]
      properties:
        originLat:
          type: number
        originLng:
          type: number
        destLat:
          type: number
        destLng:
          type: number
        destinationName:
          type: string
    SmartRouteResponse:
      type: object
      required: [mainRoute, isCongested]
      properties:
        isCongested:
          type: boolean
        mainRoute:
          $ref: "#/components/schemas/RouteOption"
        alternativeRoute:
          $ref: "#/components/schemas/RouteOption"
        pointsIfAlternative:
          type: integer
        congestionMessage:
          type: ["string", "null"]
    RouteOption:
      type: object
      required: [name, estimatedMinutes, distanceKm, waypoints]
      properties:
        name:
          type: string
        estimatedMinutes:
          type: integer
        distanceKm:
          type: number
        waypoints:
          type: array
          items:
            type: array
            items:
              type: number
        trafficLevel:
          type: string
          enum: [free, slow, congested]
        parkingZone:
          type: ["string", "null"]
...
    Obstacle:
      type: object
      required: [id, lat, lng, obstacleType, severity, isActive, createdAt]
      properties:
        id:
          type: integer
        lat:
          type: number
        lng:
          type: number
        obstacleType:
          type: string
          enum: [pothole, barrier, parked_vehicle, construction, other]
        severity:
          type: string
          enum: [low, medium, high]
        isActive:
          type: boolean
        description:
          type: string
        affectsPathId:
          type: ["integer", "null"]
        createdAt:
          type: string
        reporterUsername:
          type: ["string", "null"]
    ObstacleInput:
      type: object
      required: [lat, lng, obstacleType, severity]
      properties:
        lat:
          type: number
        lng:
          type: number
        obstacleType:
          type: string
          enum: [pothole, barrier, parked_vehicle, construction, other]
        severity:
          type: string
          enum: [low, medium, high]
        description:
          type: string
        affectsPathId:
          type: ["integer", "null"]
...
    PointsTransaction:
      type: object
      required: [id, type, points, category, description, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [earn, spend]
        points:
          type: integer
        category:
          type: string
          enum: [traffic, waste, accessibility, tourism, redemption]
        description:
          type: string
        createdAt:
          type: string
...
    RedeemInput:
      type: object
      required: [rewardId]
      properties:
        rewardId:
          type: integer
...
    AuthUser:
      type: object
      required: [id, email, firstName, lastName, profileImageUrl]
      properties:
        id:
          type: string
        email:
          type: ["string", "null"]
        firstName:
          type: ["string", "null"]
        lastName:
          type: ["string", "null"]
        profileImageUrl:
          type: ["string", "null"]
    AuthUserEnvelope:
      type: object
      required: [user]
      properties:
        user:
          oneOf:
            - $ref: '#/components/schemas/AuthUser'
            - type: "null"
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string