# lib/api-zod/src/generated/types/errorEnvelope.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Urban Up - Smart City & Tourism Platform API
 * OpenAPI spec version: 0.1.0
 */

export interface ErrorEnvelope {
  error: string;
}